requirements/base.txt
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
dj-database-url==2.2.0
psycopg2==2.9.9